+++ src/__sdk.ts
@@ -1,1 +1,1 @@
-export default '5.31.2';
+export default '5.32.3';